MD3: emulate DirectBeamEnabled feature
The MD3 have a boolean DirectBeamEnabled attribute. When DirectBeamEnabled
is set to false, MD3 will prevent moving beamstop out of beam or opening
fast shutter, if that operation will put direct beam on the detector.

Adds emulation of this feature. Adds logic to mirror the checks that
MD3 performs, taking onto consideration:

 - DirectBeamEnabled attribute value
 - FastShutterIsOpen attribute value
 - Beamstop current and requested positions

Modifies the Attribute class, so that it also takes a callback
function in the constructor. It is a mechanism that allows
performing value validation / transformation before a value is set.