BioMAX: sync-up MD3Down emulation with MD3Up
Copy and paste MicroMAX MD3Up emulation into the BioMAX MD3Down emulation.

Note, that this is probably not 100% accurate, there is probably
some nuances that need to be tweak to get the emulation accurate
with the real hardware. But, for now, this is good enought
for our MXCuBE work.